table/path: Refine Path structure
First, this patch separates the fields of Path structure into "mutable"
and "immutable" fields. The immutable fields means the unchanged fields
between the original Path instance and the cloned Path instances when
applying policies or advertising the path.

Then, this patch makes the "immutable" fields to be shared among the
original Path instance and the cloned Path instances to reduce the
memory usage.

Also, this patch introduces "pathAttrsFilter" which stores only diff
when updating or deleting the specific path attributes.

Signed-off-by: IWASE Yusuke <iwase.yusuke0@gmail.com>